feat: add custom field for Cargo Detail to allow billing on weight
This commit introduces a new custom field named "allow_bill_on_weight" to the "Cargo Detail" doctype. The field is of type "Check" and is designed to enable or disable billing based on weight. This addition enhances the functionality of the Fleet Management system by providing users with more control over billing options related to cargo handling.